Implemented ASCII bip level files loading and exporting Fixed viewport bug offsetting by 1 on game load Removed Redundant, not needed calls like viewport limiting after level load

+++ src/LevelEditorMenu.c
@@ -7,9 +7,10 @@
 #include "Pd_helperfuncs.h"
 #include "pd_api.h"
 #include "Crank.h"
+#include "LevToBip.h"
 
 char *PackName,*CreatorName;
-
+char* File1, * File2;
 void LevelEditorMenuInit()
 {
 	setCrankMoveThreshold(180);
@@ -27,6 +28,7 @@ void LevelEditorMenu()
 		LevelEditorMenuInit();
 		GameState = GSLevelEditorMenu;
 	}
+
 	int CrankMove = crankUpdate();
 	int id = -1;
 	bool response = false;
@@ -34,11 +36,11 @@ void LevelEditorMenu()
 	{
 		pd->graphics->drawBitmap(IMGTitleScreen,0,0,kBitmapUnflipped);
 		printTitleInfo();
-		pd->graphics->fillRect(40 * UI_WIDTH_SCALE, 70 * UI_HEIGHT_SCALE, 240 * UI_WIDTH_SCALE, 100 * UI_HEIGHT_SCALE, kColorWhite);
-		pd->graphics->drawRect(40 * UI_WIDTH_SCALE, 70 * UI_HEIGHT_SCALE, 240 * UI_WIDTH_SCALE, 100 * UI_HEIGHT_SCALE, kColorBlack);
-		pd->graphics->drawRect(42 * UI_WIDTH_SCALE, 72 * UI_HEIGHT_SCALE, 237 * UI_WIDTH_SCALE, 97 * UI_HEIGHT_SCALE, kColorBlack);
-		pd->system->formatString(&Tekst,"Create New Levelpack\nLoad Selected LevelPack\nDelete Selected Levelpack\n<%s>\nMain Menu",LevelPackName);
-		drawText(BigFont,Tekst,strlen(Tekst),kASCIIEncoding, 70 * UI_WIDTH_SCALE, 77 * UI_HEIGHT_SCALE);
+		pd->graphics->fillRect(40 * UI_WIDTH_SCALE, 62 * UI_HEIGHT_SCALE, 240 * UI_WIDTH_SCALE, 116 * UI_HEIGHT_SCALE, kColorWhite);
+		pd->graphics->drawRect(40 * UI_WIDTH_SCALE, 62 * UI_HEIGHT_SCALE, 240 * UI_WIDTH_SCALE, 116 * UI_HEIGHT_SCALE, kColorBlack);
+		pd->graphics->drawRect(42 * UI_WIDTH_SCALE, 64 * UI_HEIGHT_SCALE, 236 * UI_WIDTH_SCALE, 112 * UI_HEIGHT_SCALE, kColorBlack);
+		pd->system->formatString(&Tekst,"Create New Levelpack\nLoad Selected LevelPack\nDelete Selected Levelpack\nExport to Bip File\n<%s>\nMain Menu",LevelPackName);
+		drawText(BigFont, Tekst, strlen(Tekst), kASCIIEncoding, 70 * UI_WIDTH_SCALE, 69 * UI_HEIGHT_SCALE);
 		pd->system->realloc(Tekst, 0);
 		Tekst = pd->system->realloc(NULL, 100);
 		if (Selection > 1)
@@ -50,8 +52,43 @@ void LevelEditorMenu()
 		}
 		else
 			strcpy(Tekst,">>");
-		drawText(BigFont, Tekst, strlen(Tekst), kASCIIEncoding, 45 * UI_WIDTH_SCALE, 77 * UI_HEIGHT_SCALE);
+		drawText(BigFont, Tekst, strlen(Tekst), kASCIIEncoding, 45 * UI_WIDTH_SCALE, 69 * UI_HEIGHT_SCALE);
 		pd->system->realloc(Tekst, 0);
+
+		int numLeft = ConvertLevToBipContinue();
+		if (numLeft > 0)
+		{
+			pd->system->formatString(&Tekst, "Exporting levels %d remaining", numLeft);
+			//-1 is without user input required
+			AskQuestion(-1, Tekst);
+			pd->system->realloc(Tekst, 0);
+			return;
+		}
+		
+		if (numLeft == 0)
+		{
+			playMenuSelectSound();
+			pd->system->formatString(&Tekst, "Levelpack %s was exported\nto following location:\n<gamedata>/%s\n\nPress (A) to continue", LevelPackName, File2);
+			AskQuestion(IDBipExported, Tekst);
+			pd->system->realloc(Tekst, 0);
+			pd->system->realloc(File1, 0);
+			pd->system->realloc(File2, 0);
+			char* tmp;
+			pd->system->formatString(&tmp, "%s", LevelPackName);
+			SearchForLevelPacks();
+			for (Teller = 0; Teller < InstalledLevelPacksCount; Teller++)
+				if (strcmp(tmp, InstalledLevelPacks[Teller]) == 0)
+				{
+					SelectedLevelPack = Teller;
+					if (LevelPackName)
+						pd->system->realloc(LevelPackName, 0);
+					pd->system->formatString(&LevelPackName, "%s", InstalledLevelPacks[SelectedLevelPack]);
+				}
+			pd->system->realloc(tmp, 0);
+			return;
+		}
+		
+		
 		if ((currButtons & kButtonB) && !(prevButtons & kButtonB))
 		{
 			playMenuBackSound();
@@ -64,7 +101,7 @@ void LevelEditorMenu()
 				frameLeftStart = framecount;
 			if ((framecount - frameLeftStart) % MenuUpdateTicks == 0)
 			{
-				if (Selection == 4)
+				if (Selection == 5)
 					if (InstalledLevelPacksCount > 0)
 					{
 						SelectedLevelPack--;
@@ -73,6 +110,9 @@ void LevelEditorMenu()
 						if (LevelPackName)
 							pd->system->realloc(LevelPackName, 0);
 						pd->system->formatString(&LevelPackName, "%s", InstalledLevelPacks[SelectedLevelPack]);
+						pd->system->formatString(&FileName, "levelpacks/%s", LevelPackName);
+						CLevelPackFile_loadFile(LevelPackFile, FileName, NrOfCols, NrOfRows, true);
+						pd->system->realloc(FileName, 0);
 						LoadNormalCreatorName();
 						playMenuSound();
 						SaveSettings();
@@ -86,7 +126,7 @@ void LevelEditorMenu()
 				frameRightStart = framecount;
 			if ((framecount - frameRightStart) % MenuUpdateTicks == 0)
 			{
-				if (Selection == 4)
+				if (Selection == 5)
 					if (InstalledLevelPacksCount > 0)
 					{
 						SelectedLevelPack++;
@@ -95,6 +135,9 @@ void LevelEditorMenu()
 						if (LevelPackName)
 							pd->system->realloc(LevelPackName, 0);
 						pd->system->formatString(&LevelPackName, "%s", InstalledLevelPacks[SelectedLevelPack]);
+						pd->system->formatString(&FileName, "levelpacks/%s", LevelPackName);
+						CLevelPackFile_loadFile(LevelPackFile, FileName, NrOfCols, NrOfRows, true);
+						pd->system->realloc(FileName, 0);
 						LoadNormalCreatorName();
 						playMenuSound();
 						SaveSettings();
@@ -122,7 +165,7 @@ void LevelEditorMenu()
 				frameDownStart = framecount;
 			if (((framecount - frameDownStart) % MenuUpdateTicks == 0) || (CrankMove == CRANKMOVERIGHT))
 			{
-				if (Selection < 5)
+				if (Selection < 6)
 				{
 					Selection++;
 					playMenuSound();
@@ -139,12 +182,15 @@ void LevelEditorMenu()
 					PackName = GetString(1, 55*UI_WIDTH_SCALE,85*UI_HEIGHT_SCALE, "Enter the Levelpack name:", 19);										
 					break;
 				case 2:
-					if (InstalledLevelPacksCount >0)
+					if (InstalledLevelPacksCount > 0)
 					{
+						pd->system->formatString(&FileName, "levelpacks/%s", LevelPackName);
+						CLevelPackFile_loadFile(LevelPackFile, FileName, NrOfCols, NrOfRows, false);
+						pd->system->realloc(FileName, 0);
 						LoadNormalCreatorName();
 						FindLevels();
-						SelectedLevel=0;
-						GameState=GSStageSelectInit;
+						SelectedLevel = 0;
+						GameState = GSStageSelectInit;
 						playMenuSelectSound();
 					}
 					break;
@@ -158,6 +204,29 @@ void LevelEditorMenu()
 					}
 					break;
 				case 4:
+					playMenuSelectSound();
+					char* dot = strrchr(LevelPackName, '.');
+					bool canContinue = true;
+					if (dot)
+					{
+						if ((strcmp(dot, ".bip") == 0) || (strcmp(dot, ".BIP") == 0) ||
+							(strcmp(dot, ".txt") == 0) || (strcmp(dot, ".TXT") == 0))
+						{
+							canContinue = false;
+						}
+					}
+					if (canContinue)
+					{
+						//-1 is without user input required
+						AskQuestion(-1, "Exporting levels");
+						pd->system->formatString(&File1, "levelpacks/%s._lev", LevelPackName);
+						pd->system->formatString(&File2, "levelpacks/%s.bip", LevelPackName);
+						ConvertLevToBip(File1, File2, LevelPackName, NormalCreateName);
+					}
+					else
+						AskQuestion(IDBipRefuseExport, "Ascii (*.bip, *.txt) levelpacks can not be exported to bip files\nOnly completely self created levelpacks can be exported!\n\nPress (A) to continue");
+					break;
+				case 5:
 					if (InstalledLevelPacksCount > 0)
 					{
 						SelectedLevelPack++;
@@ -166,12 +235,15 @@ void LevelEditorMenu()
 						if (LevelPackName)
 							pd->system->realloc(LevelPackName, 0);
 						pd->system->formatString(&LevelPackName, "%s", InstalledLevelPacks[SelectedLevelPack]);
+						pd->system->formatString(&FileName, "levelpacks/%s", LevelPackName);
+						CLevelPackFile_loadFile(LevelPackFile, FileName, NrOfCols, NrOfRows, true);
+						pd->system->realloc(FileName, 0);
 						LoadNormalCreatorName();
 						playMenuSound();
 						SaveSettings();
 					}
 					break;
-				case 5:
+				case 6:
 					playMenuSelectSound();
 					GameState = GSTitleScreenInit;
 					break;
@@ -201,7 +273,7 @@ void LevelEditorMenu()
 						SDFile* Fp = pd->file->open(FileName, kFileWrite);
 						if (Fp)
 						{
-							pd->file->write(Fp, CreatorName, strlen(CreatorName)+1);
+							pd->file->write(Fp, CreatorName, (unsigned int)(strlen(CreatorName) + 1));
 							pd->file->close(Fp);
 						}
 						pd->system->realloc(FileName, 0);
@@ -218,6 +290,13 @@ void LevelEditorMenu()
 				}
 				pd->system->realloc(PackName, 0);
 				pd->system->realloc(CreatorName, 0);
+				if (InstalledLevelPacksCount > 0)
+				{
+					pd->system->formatString(&FileName, "levelpacks/%s", LevelPackName);
+					CLevelPackFile_loadFile(LevelPackFile, FileName, NrOfCols, NrOfRows, true);
+					pd->system->realloc(FileName, 0);
+					LoadNormalCreatorName();
+				}
 				LoadNormalCreatorName();
 			}
 		}
@@ -246,13 +325,28 @@ void LevelEditorMenu()
 		{
 			if(id == IDDeleteLevelPack)
 			{
-				if(response)
+				if (response)
 				{
 					FindLevels();
-					pd->system->formatString(&Tekst,"levelpacks/%s._lev", LevelPackName);
-					pd->file->unlink(Tekst,true);
+					char* dot = strrchr(LevelPackName, '.');
+					if (dot && ((strcmp(dot, ".bip") == 0) || (strcmp(dot, ".BIP") == 0) ||
+						(strcmp(dot, ".txt") == 0) || (strcmp(dot, ".TXT") == 0)))
+					{
+						pd->system->formatString(&Tekst, "levelpacks/%s", LevelPackName);
+						pd->file->unlink(Tekst, true);
+					}
+					
+					pd->system->formatString(&Tekst, "levelpacks/%s._lev", LevelPackName);
+					pd->file->unlink(Tekst, true);
 					pd->system->realloc(Tekst, 0);
 					SearchForLevelPacks();
+					if (InstalledLevelPacksCount > 0)
+					{
+						pd->system->formatString(&FileName, "levelpacks/%s", LevelPackName);
+						CLevelPackFile_loadFile(LevelPackFile, FileName, NrOfCols, NrOfRows, true);
+						pd->system->realloc(FileName, 0);
+						LoadNormalCreatorName();
+					}
 					LoadNormalCreatorName();
 				}
 			}
